fix: code cleanup & fix units
